backend/requirements.txt
flask
flask-cors
python-dotenv
pymongo
passlib
pyjwt
gunicorn
python-multipart
requests
numpy
ultralytics
SpeechRecognition
pydub
sentencepiece
protobuf
google-generativeai